Entity iterator test and fixes
